+++ build_modules/CHANGELOG.md
@@ -1,7 +1,13 @@
 ## 2.0.0
 
+### Breaking Changes
+
 - Remove the `merge` method from `Module` and replace with a static
   `Module.merge`. Module instances are now immutable.
+- Remove `jsId`, and `jsSourceMapId` from `Module`.
+
+### Improvements
+
 - Update the kernel worker to pass input digests, along with
   `--reuse-compiler-result` and `--use-incremental-compiler`.